Game

Documentation

data Player

Data fields required | Player returns which player is on the cell

Constructors

Player1

Player2

Dot

data Cell

Cell shows the state of the cell

Constructors

Empty

Full Player

data State

State is the state of teh game

Constructors

Menu

Running

RunningB

GameOver (Maybe Player)

type Board = Array (Int, Int) Cell

Board is for storing the 2-D array

data Choice

Choice is the choice the user selects in the menu

Constructors

Multi

Single

type MenuBoard = Array (Int, Int) Choice

MenuBoard is the array to represent the menu board

data Game

Game returns the current state of teh game witn all it's member variables

Constructors

Game

Fields

gameBoard :: Board

menuBoard :: MenuBoard

gamePlayer :: Player

gameState :: State

player1Stone :: Int

player2Stone :: Int

maxStone1 :: Int

maxStone2 :: Int

removeStone1 :: Int

removeStone2 :: Int

moveCoords :: (Int, Int)

movedCoordSet :: Int

gameList :: [[Int]]

checkListH :: [Int]

checkListV :: [Int]

checker :: Int

playerMovedCoords :: (Int, Int)

botCoords :: (Int, Int)

n :: Int

n is the size of the grid

listForCheck :: [[Int]]

This are the list required for checking when a player forms a morris listForCheck, listForHorizontalCheck, listForVerticalCheck

listForHorizontalCheck :: [Int]

listForVerticalCheck :: [Int]

screenWidth :: Int

Screen window discription screenWidth = 644 screenHeight = 644

screenHeight :: Int

cellWidth :: Float

cellHeight :: Float

initialGame :: Game

initialGame Initial constructor for game