Logic1

Documentation

isCoordCorrect :: (Int, Int) -> Bool

The isCoordCorrect function checks whether the given coordinates are correct or not

twoHCoord :: [(Int, Int)]

twoHCoord List of coordinates which have two horizontal neighbours

twoVCoord :: [(Int, Int)]

twoVCoord | List of coordinates which have two verticle neighbours

oneHCoord :: [(Int, Int)]

oneHCoord List of coordinates which have one horizontal neighbours

oneVCoord :: [(Int, Int)]

oneVCoord List of coordinates which have one verticle neighbours

switchPlayer :: Game -> Game

switchPlayer Fuction which switches the player according to the value of checker data in the game data set

switchPlayer1 :: Game -> Game

switchPlaye1 No checker invollved, this function switches the player when called

mousePosAsMenuCellCoord :: (Float, Float) -> (Int, Int)

mousePosAsMenuCellCoord This funciton converts the mouse coordinates to menu cell coordinates

mousePosAsCellCoord :: (Float, Float) -> (Int, Int)

mousePosAsCellCoord This function converts the mouse coordinates to cell coordinates

checkGameOver :: Game -> Game

checkGameOver This funcion implements the gameover condition for the game 1) When either of a player has less than 3 stones left 2) A player does not have a viable move left

checkBoardForViable :: Game -> Int

checkBoardForViable This function checks the viability of the board, that a player still has a valid move or not

traverseBoard :: Game -> (Int, Int) -> Int

playerTurn :: Game -> (Int, Int) -> Game

playerTurn This function is for Multi Player deceiding the player turn, This is the main heart of the program this function implements all the required logic for the game

playerTurnB :: Game -> (Int, Int) -> Game

playerTurnB This function is for Single Player (when playing against a bot) deceiding the player turn,

validCellCoords :: Game -> (Int, Int) -> (Int, Int) -> Bool

validCellCoords This function checks for the validity of the corrdinates when selected to move

isUpNeighbour :: Game -> (Int, Int) -> (Int, Int) -> Int

Helper functions for checking the valid cell coordinates | isUpNeighbour This funciton traverse in one direction up down left or right and checks that the given coordinates have free neighbours

isDownNeighbour :: Game -> (Int, Int) -> (Int, Int) -> Int

isLeftNeighbour :: Game -> (Int, Int) -> (Int, Int) -> Int

isRightNeighbour :: Game -> (Int, Int) -> (Int, Int) -> Int

setCoords :: Game -> (Int, Int) -> Int -> Game

setCoords Setting the first clicked coords

checkNeighbour :: Game -> (Int, Int) -> Int

isUp :: Game -> (Int, Int) -> Int

isUp This function return the required data for movement as the neighbous position or if it is present or not

isDown :: Game -> (Int, Int) -> Int

isLeft :: Game -> (Int, Int) -> Int

isRight :: Game -> (Int, Int) -> Int

isUpB :: Game -> (Int, Int) -> (Int, Int)

isDownB :: Game -> (Int, Int) -> (Int, Int)

isLeftB :: Game -> (Int, Int) -> (Int, Int)

isRightB :: Game -> (Int, Int) -> (Int, Int)

isUpBB :: Game -> (Int, Int) -> (Int, Int)

isDownBB :: Game -> (Int, Int) -> (Int, Int)

isLeftBB :: Game -> (Int, Int) -> (Int, Int)

isRightBB :: Game -> (Int, Int) -> (Int, Int)

isRightTwoB :: Game -> (Int, Int) -> (Int, Int)

Function checks that are there and two in a line or not isRightTwoB, isLeftTwoB, isDownTwoB, isUpTwoB, isVOneB, isHOneB

isLeftTwoB :: Game -> (Int, Int) -> (Int, Int)

isDownTwoB :: Game -> (Int, Int) -> (Int, Int)

isUpTwoB :: Game -> (Int, Int) -> (Int, Int)

isVOneB :: Game -> (Int, Int) -> (Int, Int)

isHOneB :: Game -> (Int, Int) -> (Int, Int)

transformGame :: Event -> Game -> Game

transformGame This function recognise the mouse position and converts it into the cell position where the user clicks transformGame :: Game -> Game

openGame :: Game -> (Int, Int) -> Game

openGame This function is used in displaying the menu

horizontalLine :: Game -> [Int] -> Cell

Placing logic :: -> when some one places a stone check for that its movement has created a morris or not -> the horiontal line funcion when mapped creates a list showing that are there any horizontal morris or not -> same is done by verticle -> as the game rules once a morris is formed that norris gets blocked until some change or movement are made in that morris -> so there are take other functions which checks that the morris formed is a new one or the old one -> if the morris is a new one than you give the current player a second chance to take opposite players stone -> this locking and unloccking thing is mainted by the list called checklistH and checklistV -> there are funcions called remover which removes the stone of oppostie layer and blocks the list by which the current player got chance to take -> similarly there are unblock list funciton which unblocks the morris list when there are any changes or movement in the configuration

verticleLine :: Game -> [Int] -> Cell

horizontalLineUnblock :: Game -> [Int] -> Cell

verticleLineUnblock :: Game -> [Int] -> Cell

finalHorizontalCheck :: Game -> [Cell]

Mapping above function to create a list finalHorizontalCheck, finalVerticalCheck

finalVerticalCheck :: Game -> [Cell]

finalHorizontalCheckUnblocker :: Game -> [Cell]

finalVerticalCheckUnbloccker :: Game -> [Cell]

takeOtherH :: Game -> Int -> Int

Blocker functions takeOtherH, takeOtherV, takeOther

takeOtherV :: Game -> Int -> Int

takeOther :: Game -> Int

unblockOtherH :: Game -> Int -> Int

Unblocker funcitons unblockOtherH, unblockOtherV, listUnblockerH, listUnblockerV

unblockOtherV :: Game -> Int -> Int

listUnblockerH :: Game -> Game

listUnblockerV :: Game -> Game

replaceNth :: Int -> [Int] -> Int -> [Int]

replaceNth Funciton to replace nth postion of list with new value

removerH :: (Int, Int) -> Game -> Game

removerH, removerV Funcitons for removing others stone and blocking the list

removerV :: (Int, Int) -> Game -> Game

removerHB :: Game -> Game

These are for bot removerHB, removerVB

removerVB :: Game -> Game

playerSwitcherConfirm :: Game -> Game

botMoveThree :: Game -> (Int, Int) -> (Int, Int)

Movement Logic :: -> First we store the corrdinates of the click position by player in a variable -> then we check wheter the coordinates are coorect or not -> if correct we let him to click to other neighbour -> if that are also correct we just move -> bot also follows same logic but it does both step in one -> it first checks if he can make or block a morris -> if not than it tries to block the player

botMoveThree This Function botMove to check that this empty postion is where a morris can be formed

simpleTraverseBotMove :: Game -> (Int, Int) -> (Int, Int)

simpleTraverseBotMove Find the first position on board for bot for which it can move

posneighSimpTrv :: Game -> (Int, Int) -> (Int, Int)

posneighSimpTrv Movement to its neighbour

botMove :: Game -> (Int, Int) -> Game

botMove Main function to handle bot movement

botMoveThreeFly :: Game -> (Int, Int) -> (Int, Int)

Move When Less than 3 stones -> when a bot or player has less than three stones than it can move its stone to anywhere on board -> so the locgic is implemented here

botMoveThreeFly Fly to the corrdinate were morris can be formed or blocked

botFlyMove :: Game -> (Int, Int) -> Game

botFlyMove Checking bot fly movement

traverseBoardBot :: Game -> (Int, Int) -> Game

traverseBoardBot Traverse board for simple board movement

threeTakerH :: Game -> Int -> (Int, Int)

threeTakerV :: Game -> Int -> (Int, Int)

traverseBoardBotTake :: Game -> (Int, Int) -> (Int, Int)

traverseBoardBotMove :: Game -> (Int, Int) -> (Int, Int)

twoInRowChecker :: Game -> (Int, Int) -> Game

twoInRowChecker This function checks are there any two same placed stone in a row

twoInRowCheckerBotMove :: Game -> (Int, Int) -> (Int, Int)

twoInRowCheckerBotMove To take action accordin to two stones in a row

botMoveOnPlayer :: Game -> Game

botMoveOnPlayer The function which tries to block the player when no morris can be formed or blocked