Rendering

Documentation

player1Color :: Color

Intial colors for the different graphics

player2Color :: Color

gridLightPlayer1 :: Color

gridLightPlayer2 :: Color

blackcolor :: Color

whitecolor :: Color

tieColor :: Color

boardGrid :: Picture

boardGrid defines the lines on the board

menuBoardLine :: Picture

menuBoardLine defines the lines on the menu board

rectangleMenuBlack :: Picture

tileRenderMulti :: Picture

tileRenderMulti returns the Picture from the string for Multi cell part

tileRenderSingle :: Picture

tileRenderSingle returns the Picture from the string for Single cell part

boardMenuPicture :: Game -> MenuBoard -> Picture

boardAsRunningPicture :: Game -> Board -> Picture

boardAsRunningPicture renders the picture depending on the board provided

snapPictureToCell :: Picture -> (Int, Int) -> Picture

snapPictureToCell places the picture back to the cell

snapPictureToMenuCell :: Picture -> (Float, Float) -> Picture

aCell :: Picture

aCell returns a solid circle picture with a defined radius of Player 1

bCell :: Picture

bCell returns a solid circle picture with a defined radius of Player 2

dotCell :: Picture

dotCell returns a solid circle picture with a smaller defined radius for no players

cellsOfBoard :: Board -> Cell -> Picture -> Picture

cellsOfBoard maps the picture to the cells according to the cell type

menuCellsBoard :: MenuBoard -> Choice -> Picture -> Picture

menuCellsBoard is in accordance to the cellsOfBoard function

tileRenderMultiCell :: MenuBoard -> Picture

tileRenderSingleCell :: MenuBoard -> Picture

aCellsOfBoard :: Board -> Picture

aCellsOfBoard returns the cells where the cells are of a type (Full Player1)

bCellsOfBoard :: Board -> Picture

bCellsOfBoard returns the cells where the cells are of b type (Full Player2)

dotCellsOfBoard :: Board -> Picture

dotCellsOfBoard returns the cells where the cells are of dot type (Full Dot)

gameAsPicture :: Game -> Picture

gameAsPicture renders the board or the menu screen depending on the Game state